!  make kafka-logback-appender generic in the event type to support alternative events like logback-access (#24). Rename PatternLayoutKafkaMessageEncoder to LayoutKafkaMessageEncoder (with deprecated class alias) (#9). Add javadoc since hint. Switch to 0.1.0-SNAPSHOT version.
