+++ src/main/java/com/github/danielwegener/logback/kafka/encoding/PatternLayoutKafkaMessageEncoder.java
@@ -1,56 +1,19 @@
 package com.github.danielwegener.logback.kafka.encoding;
 
-import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.core.Layout;
 
 import java.nio.charset.Charset;
 
 /**
- * A KafkaMessageEncoder that can be configured with a {@link Layout} and a {@link Charset} and creates
- * a serialized string for each logging event using the given layout.
+ * @deprecated Use LayoutKafkaMessageEncoder instead!
+ * @since 0.0.1
  */
-public class PatternLayoutKafkaMessageEncoder extends KafkaMessageEncoderBase<ILoggingEvent> {
+@Deprecated
+public class PatternLayoutKafkaMessageEncoder<E> extends LayoutKafkaMessageEncoder<E> {
 
-    public PatternLayoutKafkaMessageEncoder() {
-    }
+    public PatternLayoutKafkaMessageEncoder() {}
 
-    public PatternLayoutKafkaMessageEncoder(Layout<ILoggingEvent> layout, Charset charset) {
-        this.layout = layout;
-        this.charset = charset;
-    }
-
-    private Layout<ILoggingEvent> layout;
-    private Charset charset;
-    private static final Charset UTF8 = Charset.forName("UTF-8");
-
-    @Override
-    public void start() {
-        if (charset == null) {
-            addInfo("No charset specified for PatternLayoutKafkaEncoder. Using default UTF8 encoding.");
-            charset = UTF8;
-        }
-        super.start();
-    }
-
-    @Override
-    public byte[] doEncode(ILoggingEvent loggingEvent) {
-        final String message = layout.doLayout(loggingEvent);
-        return message.getBytes(charset);
-    }
-
-    public void setLayout(Layout<ILoggingEvent> layout) {
-        this.layout = layout;
-    }
-
-    public void setCharset(Charset charset) {
-        this.charset = charset;
-    }
-
-    public Layout<ILoggingEvent> getLayout() {
-        return layout;
-    }
-
-    public Charset getCharset() {
-        return charset;
+    public PatternLayoutKafkaMessageEncoder(Layout<E> layout, Charset charset) {
+        super(layout, charset);
     }
 }
