+++ src/main/java/com/github/danielwegener/logback/kafka/keying/RoundRobinKeyingStrategy.java
@@ -1,16 +1,14 @@
 package com.github.danielwegener.logback.kafka.keying;
 
-
-import ch.qos.logback.classic.spi.ILoggingEvent;
-
 /**
  * Evenly distributes all written log messages over all available kafka partitions.
  * This strategy can lead to unexpected read orders on clients.
+ * @since 0.0.1
  */
-public class RoundRobinKeyingStrategy implements KeyingStrategy {
+public class RoundRobinKeyingStrategy implements KeyingStrategy<Object> {
 
     @Override
-    public byte[] createKey(ILoggingEvent e) {
+    public byte[] createKey(Object e) {
         return null;
     }
 }
